[Done user section

+++ src/app/modules/user/user.service.ts
@@ -1,25 +1,27 @@
-import { User, Prisma } from '@prisma/client';
+import { User, Prisma, userRole } from '@prisma/client';
 import { paginationHelpers } from '../../../helpers/paginationHelper';
 import { IGenericResponse } from '../../../interfaces/common';
 import { IPaginationOptions } from '../../../interfaces/pagination';
 import prisma from '../../../shared/prisma';
 import { userSearchableFields } from './user.constants';
 import { IUserFilterRequest } from './user.interface';
+import ApiError from '../../../errors/ApiError';
+import httpStatus from 'http-status';
 
 const getAllUsers = async (
   filterOptions: IUserFilterRequest,
   paginationOptions: IPaginationOptions
-): Promise<IGenericResponse<User[]>> => {
-  const { search, ...filtersData } = filterOptions;
+): Promise<IGenericResponse<(User[] | any) | null>> => {
+  const { searchTerm, ...filtersData } = filterOptions;
 
   const { page, limit, skip, sortBy, sortOrder } =
     paginationHelpers.calculatePagination(paginationOptions);
 
   const andConditions = [];
-  if (search) {
+  if (searchTerm) {
     andConditions.push({
       OR: userSearchableFields.map(field => ({
-        [field]: { contains: search, mode: 'insensitive' },
+        [field]: { contains: searchTerm, mode: 'insensitive' },
       })),
     });
   }
@@ -37,6 +39,15 @@ const getAllUsers = async (
 
   const result = await prisma.user.findMany({
     where: whereConditions,
+    select: {
+      id: true,
+      email: true,
+      role: true,
+      Admin: true,
+      Customer: true,
+      createdAt: true,
+      updatedAt: true,
+    },
     skip: skip,
     take: limit,
     orderBy:
@@ -63,10 +74,19 @@ const getAllUsers = async (
   };
 };
 
-const getSingleUser = async (id: string): Promise<User | null> => {
+const getSingleUser = async (id: string): Promise<(User | any) | null> => {
   const result = await prisma.user.findUnique({
     where: {
       id,
+    },
+    select: {
+      id: true,
+      email: true,
+      role: true,
+      Admin: true,
+      Customer: true,
+      createdAt: true,
+      updatedAt: true,
     },
   });
   return result;
@@ -86,12 +106,56 @@ const updateUser = async (
 };
 
 const deleteUser = async (id: string): Promise<User | null> => {
-  const result = await prisma.user.delete({
-    where: {
-      id,
-    },
+  const deletedUser = await prisma.$transaction(async transactionClient => {
+    let result;
+    const existingUser = await transactionClient.user.findUnique({
+      where: {
+        id,
+      },
+      include: {
+        Admin: true,
+        Customer: true,
+      },
+    });
+
+    console.log({ existingUser });
+
+    if (existingUser) {
+      const role = existingUser.role;
+      let isDelete;
+      if (role === userRole.super_admin) {
+        throw new ApiError(
+          httpStatus.BAD_REQUEST,
+          'super admin can not deleted'
+        );
+      } else if (role === userRole.admin) {
+        isDelete = await transactionClient.admin.delete({
+          where: {
+            id: existingUser.Admin[0].id,
+          },
+        });
+      } else {
+        isDelete = await transactionClient.customer.delete({
+          where: {
+            id: existingUser.Customer[0].id,
+          },
+        });
+      }
+
+      if (isDelete) {
+        result = await prisma.user.delete({
+          where: {
+            id,
+          },
+        });
+      }
+    }
+
+    return result;
   });
-  return result;
+
+  if (deletedUser) return deletedUser;
+  throw new ApiError(httpStatus.BAD_REQUEST, 'Unable to delete user ');
 };
 
 export const UserService = {
